rename enclave mod to nitro 🔋
