+++ crates/ev-enclave/src/describe/mod.rs
@@ -2,7 +2,7 @@ pub mod error;
 
 use crate::common::resolve_output_path;
 use crate::docker::{error::DockerError, utils::verify_docker_is_running};
-use crate::enclave;
+use crate::nitro;
 use crate::progress::get_tracker;
 use error::DescribeError;
 
@@ -10,7 +10,7 @@ pub fn describe_eif(
     eif_path: &str,
     verbose: bool,
     no_cache: bool,
-) -> Result<enclave::DescribeEif, DescribeError> {
+) -> Result<nitro::DescribeEif, DescribeError> {
     let eif_path = std::path::Path::new(eif_path);
     if !eif_path.exists() {
         return Err(DescribeError::EIFNotFound(eif_path.to_path_buf()));
@@ -26,9 +26,9 @@ pub fn describe_eif(
 
     let supplied_path: Option<&str> = None;
     let output_path = resolve_output_path(supplied_path).unwrap();
-    enclave::build_nitro_cli_image(output_path.path(), None, verbose, no_cache)?;
+    nitro::build_nitro_cli_image(output_path.path(), None, verbose, no_cache)?;
 
-    let description = enclave::describe_eif(&absolute_path, verbose)?;
+    let description = nitro::describe_eif(&absolute_path, verbose)?;
     describe_progress.finish_with_message("PCRs retrieved.");
 
     Ok(description)
