+++ crates/ev-enclave/src/lib.rs
@@ -12,13 +12,13 @@ pub mod describe;
 #[cfg(feature = "internal_dependency")]
 pub mod dev;
 pub mod docker;
-pub mod enclave;
 #[cfg(feature = "internal_dependency")]
 pub mod encrypt;
 #[cfg(feature = "internal_dependency")]
 pub mod env;
 pub mod logs;
 pub mod migrate;
+pub mod nitro;
 pub mod progress;
 pub mod restart;
 mod version;
