+++ crates/ev-enclave/src/nitro/nitro-cli-image.Dockerfile
